Move the Globalization and Encoding contracts to System.Runtime
This change is to add the globalization and encoding to System.Runtime and update
System.Globalization.* and System.Text.Encoding contarcts too just type forward instead of
having any implementation
