+++ BuildToolsVersion.txt
@@ -1,1 +1,1 @@
-1.0.26-prerelease-00723-02
+1.0.26-prerelease-00729-01
