+++ src/System.Globalization.Calendars/pkg/aot/System.Globalization.Calendars.pkgproj
@@ -7,11 +7,6 @@
     <PreventImplementationReference>true</PreventImplementationReference>
   </PropertyGroup>
 
-  <ItemGroup>
-    <ProjectReference Include="..\..\src\System.Globalization.Calendars.csproj">
-      <TargetGroup>netcore50aot</TargetGroup>
-    </ProjectReference>
-  </ItemGroup>
 
   <Import Project="$([MSBuild]::GetDirectoryNameOfFileAbove($(MSBuildThisFileDirectory), dir.targets))\dir.targets" />
 </Project>
