+++ src/System.Globalization.Calendars/ref/System.Globalization.Calendars.Forwards.cs
@@ -1,0 +1,25 @@
+// Licensed to the .NET Foundation under one or more agreements.
+// The .NET Foundation licenses this file to you under the MIT license.
+// See the LICENSE file in the project root for more information.
+// ------------------------------------------------------------------------------
+// Changes to this file must follow the http://aka.ms/api-review process.
+// ------------------------------------------------------------------------------
+
+[assembly: System.Runtime.CompilerServices.TypeForwardedTo(typeof(System.Globalization.ChineseLunisolarCalendar))]
+[assembly: System.Runtime.CompilerServices.TypeForwardedTo(typeof(System.Globalization.EastAsianLunisolarCalendar))]
+[assembly: System.Runtime.CompilerServices.TypeForwardedTo(typeof(System.Globalization.GregorianCalendar))]
+[assembly: System.Runtime.CompilerServices.TypeForwardedTo(typeof(System.Globalization.GregorianCalendarTypes))]
+[assembly: System.Runtime.CompilerServices.TypeForwardedTo(typeof(System.Globalization.HebrewCalendar))]
+[assembly: System.Runtime.CompilerServices.TypeForwardedTo(typeof(System.Globalization.HijriCalendar))]
+[assembly: System.Runtime.CompilerServices.TypeForwardedTo(typeof(System.Globalization.JapaneseCalendar))]
+[assembly: System.Runtime.CompilerServices.TypeForwardedTo(typeof(System.Globalization.JapaneseLunisolarCalendar))]
+[assembly: System.Runtime.CompilerServices.TypeForwardedTo(typeof(System.Globalization.JulianCalendar))]
+[assembly: System.Runtime.CompilerServices.TypeForwardedTo(typeof(System.Globalization.KoreanCalendar))]
+[assembly: System.Runtime.CompilerServices.TypeForwardedTo(typeof(System.Globalization.KoreanLunisolarCalendar))]
+[assembly: System.Runtime.CompilerServices.TypeForwardedTo(typeof(System.Globalization.PersianCalendar))]
+[assembly: System.Runtime.CompilerServices.TypeForwardedTo(typeof(System.Globalization.TaiwanCalendar))]
+[assembly: System.Runtime.CompilerServices.TypeForwardedTo(typeof(System.Globalization.TaiwanLunisolarCalendar))]
+[assembly: System.Runtime.CompilerServices.TypeForwardedTo(typeof(System.Globalization.ThaiBuddhistCalendar))]
+[assembly: System.Runtime.CompilerServices.TypeForwardedTo(typeof(System.Globalization.UmAlQuraCalendar))]
+
+
