+++ src/System.Globalization.Calendars/src/System.Globalization.Calendars.builds
@@ -4,13 +4,7 @@
   <ItemGroup>
     <Project Include="System.Globalization.Calendars.csproj" />
     <Project Include="System.Globalization.Calendars.csproj">
-      <TargetGroup>net46</TargetGroup>
-    </Project>
-    <Project Include="System.Globalization.Calendars.csproj">
-      <TargetGroup>netcore50</TargetGroup>
-    </Project>
-    <Project Include="System.Globalization.Calendars.csproj">
-      <TargetGroup>netcore50aot</TargetGroup>
+      <TargetGroup>net463</TargetGroup>
     </Project>
   </ItemGroup>
   <Import Project="$([MSBuild]::GetDirectoryNameOfFileAbove($(MSBuildThisFileDirectory), dir.traversal.targets))\dir.traversal.targets" />
