+++ src/System.Globalization.Calendars/src/project.json
@@ -1,5 +1,10 @@
 {
   "frameworks": {
+    "netstandard1.7": {
+      "dependencies": {
+        "Microsoft.TargetingPack.Private.CoreCLR": "1.1.0-beta-24426-03"
+      }
+    },
     "netstandard1.3": {
       "dependencies": {
         "Microsoft.TargetingPack.Private.CoreCLR": "1.1.0-beta-24426-03"
@@ -13,7 +18,7 @@
         "Microsoft.TargetingPack.Private.CoreCLR": "1.1.0-beta-24426-03"
       }
     },
-    "net46": {
+    "net463": {
       "dependencies": {
         "Microsoft.TargetingPack.NETFramework.v4.6": "1.0.1"
       }
