--- src/System.Globalization/ref/System.Globalization.cs
@@ -1,57 +1,0 @@
-// Licensed to the .NET Foundation under one or more agreements.
-// The .NET Foundation licenses this file to you under the MIT license.
-// See the LICENSE file in the project root for more information.
-// ------------------------------------------------------------------------------
-// Changes to this file must follow the http://aka.ms/api-review process.
-// ------------------------------------------------------------------------------
-
-
-namespace System.Globalization
-{
-    public static partial class CharUnicodeInfo
-    {
-        public static double GetNumericValue(char ch) { return default(double); }
-        public static double GetNumericValue(string s, int index) { return default(double); }
-        public static System.Globalization.UnicodeCategory GetUnicodeCategory(char ch) { return default(System.Globalization.UnicodeCategory); }
-        public static System.Globalization.UnicodeCategory GetUnicodeCategory(string s, int index) { return default(System.Globalization.UnicodeCategory); }
-    }
-    public partial class RegionInfo
-    {
-        public RegionInfo(string name) { }
-        public virtual string CurrencySymbol { get { return default(string); } }
-        public static System.Globalization.RegionInfo CurrentRegion { get { return default(System.Globalization.RegionInfo); } }
-        public virtual string DisplayName { get { return default(string); } }
-        public virtual string EnglishName { get { return default(string); } }
-        public virtual bool IsMetric { get { return default(bool); } }
-        public virtual string ISOCurrencySymbol { get { return default(string); } }
-        public virtual string Name { get { return default(string); } }
-        public virtual string NativeName { get { return default(string); } }
-        public virtual string TwoLetterISORegionName { get { return default(string); } }
-        public override bool Equals(object value) { return default(bool); }
-        public override int GetHashCode() { return default(int); }
-        public override string ToString() { return default(string); }
-    }
-    public partial class StringInfo
-    {
-        public StringInfo() { }
-        public StringInfo(string value) { }
-        public int LengthInTextElements { get { return default(int); } }
-        public string String { get { return default(string); } set { } }
-        public override bool Equals(object value) { return default(bool); }
-        public override int GetHashCode() { return default(int); }
-        public static string GetNextTextElement(string str) { return default(string); }
-        public static string GetNextTextElement(string str, int index) { return default(string); }
-        public static System.Globalization.TextElementEnumerator GetTextElementEnumerator(string str) { return default(System.Globalization.TextElementEnumerator); }
-        public static System.Globalization.TextElementEnumerator GetTextElementEnumerator(string str, int index) { return default(System.Globalization.TextElementEnumerator); }
-        public static int[] ParseCombiningCharacters(string str) { return default(int[]); }
-    }
-    public partial class TextElementEnumerator : System.Collections.IEnumerator
-    {
-        internal TextElementEnumerator() { }
-        public object Current { get { return default(object); } }
-        public int ElementIndex { get { return default(int); } }
-        public string GetTextElement() { return default(string); }
-        public bool MoveNext() { return default(bool); }
-        public void Reset() { }
-    }
-}
