+++ src/System.Globalization/ref/System.Globalization.csproj
@@ -8,7 +8,10 @@
     <NuGetTargetMoniker>.NETStandard,Version=v1.7</NuGetTargetMoniker>
   </PropertyGroup>
   <ItemGroup>
-    <Compile Include="System.Globalization.cs" />
+    <!-- Temporary till we publish System.Runtime updated package -->
+    <ProjectReference Include="..\..\System.Runtime\ref\System.Runtime.csproj" />
+  </ItemGroup>
+  <ItemGroup>
     <Compile Include="System.Globalization.Forwards.cs" />
   </ItemGroup>
   <ItemGroup>
