+++ src/System.Runtime.Extensions/tests/System.Runtime.Extensions.Tests.csproj
@@ -101,6 +101,8 @@
     </ProjectReference>
     <ProjectReference Include="..\..\System.Runtime\pkg\System.Runtime.pkgproj">
     </ProjectReference>
+    <ProjectReference Include="..\..\System.Text.Encoding\pkg\System.Text.Encoding.pkgproj">
+    </ProjectReference>
     <ProjectReference Include="..\..\System.Private.Uri\pkg\System.Private.Uri.pkgproj">
     </ProjectReference>
     <ProjectReference Include="$(CommonTestPath)\System\Diagnostics\RemoteExecutorConsoleApp\RemoteExecutorConsoleApp.csproj">
