+++ src/System.Text.Encoding/pkg/aot/System.Text.Encoding.pkgproj
@@ -7,11 +7,5 @@
     <PreventImplementationReference>true</PreventImplementationReference>
   </PropertyGroup>
 
-  <ItemGroup>
-    <ProjectReference Include="..\..\src\System.Text.Encoding.csproj">
-      <TargetGroup>netcore50aot</TargetGroup>
-    </ProjectReference>
-  </ItemGroup>
-
   <Import Project="$([MSBuild]::GetDirectoryNameOfFileAbove($(MSBuildThisFileDirectory), dir.targets))\dir.targets" />
 </Project>
