+++ src/System.Text.Encoding/ref/System.Text.Encoding.Forwards.cs
@@ -1,0 +1,21 @@
+// Licensed to the .NET Foundation under one or more agreements.
+// The .NET Foundation licenses this file to you under the MIT license.
+// See the LICENSE file in the project root for more information.
+// ------------------------------------------------------------------------------
+// Changes to this file must follow the http://aka.ms/api-review process.
+// ------------------------------------------------------------------------------
+
+[assembly: System.Runtime.CompilerServices.TypeForwardedTo(typeof(System.Text.Decoder))]
+[assembly: System.Runtime.CompilerServices.TypeForwardedTo(typeof(System.Text.DecoderExceptionFallback))]
+[assembly: System.Runtime.CompilerServices.TypeForwardedTo(typeof(System.Text.DecoderFallback))]
+[assembly: System.Runtime.CompilerServices.TypeForwardedTo(typeof(System.Text.DecoderFallbackBuffer))]
+[assembly: System.Runtime.CompilerServices.TypeForwardedTo(typeof(System.Text.DecoderFallbackException))]
+[assembly: System.Runtime.CompilerServices.TypeForwardedTo(typeof(System.Text.DecoderReplacementFallback))]
+[assembly: System.Runtime.CompilerServices.TypeForwardedTo(typeof(System.Text.Encoder))]
+[assembly: System.Runtime.CompilerServices.TypeForwardedTo(typeof(System.Text.EncoderExceptionFallback))]
+[assembly: System.Runtime.CompilerServices.TypeForwardedTo(typeof(System.Text.EncoderFallback))]
+[assembly: System.Runtime.CompilerServices.TypeForwardedTo(typeof(System.Text.EncoderFallbackBuffer))]
+[assembly: System.Runtime.CompilerServices.TypeForwardedTo(typeof(System.Text.EncoderFallbackException))]
+[assembly: System.Runtime.CompilerServices.TypeForwardedTo(typeof(System.Text.EncoderReplacementFallback))]
+[assembly: System.Runtime.CompilerServices.TypeForwardedTo(typeof(System.Text.Encoding))]
+[assembly: System.Runtime.CompilerServices.TypeForwardedTo(typeof(System.Text.EncodingProvider))]
