+++ src/System.Text.Encoding/ref/project.json
@@ -1,9 +1,9 @@
 {
   "dependencies": {
-    "System.Runtime": "4.0.0"
+    "System.Runtime": "4.2.0-beta-devapi-24424-01"
   },
   "frameworks": {
-    "netstandard1.3": {
+    "netstandard1.7": {
       "imports": [
         "dotnet5.4"
       ]
