+++ src/System.Text.Encoding/src/project.json
@@ -1,5 +1,13 @@
 {
   "frameworks": {
+    "netstandard1.7": {
+      "dependencies": {
+        "Microsoft.TargetingPack.Private.CoreCLR": "1.1.0-beta-24426-03"
+      },
+      "imports": [
+        "dotnet5.4"
+      ]
+    },
     "netstandard1.3": {
       "dependencies": {
         "Microsoft.TargetingPack.Private.CoreCLR": "1.1.0-beta-24426-03"
@@ -13,7 +21,7 @@
         "Microsoft.TargetingPack.Private.CoreCLR": "1.1.0-beta-24426-03"
       }
     },
-    "net46": {
+    "net463": {
       "dependencies": {
         "Microsoft.TargetingPack.NETFramework.v4.6": "1.0.1"
       }
